cleanup/#109: - remove abstract bibiserv dependencies - remove RuleBuilde for extendes cmdline options - rename DefaultPropertiesFile to ConfigurationFile - rename CommandLineValidator to Validator (the same for cloud specific implementation) - rewrite Startup to support only basic cmdline options + configurations files
