+++ bibigrid-aws/src/main/java/de/unibi/cebitec/bibigrid/aws/ConfigurationAWS.java
@@ -2,8 +2,10 @@ package de.unibi.cebitec.bibigrid.aws;
 
 import de.unibi.cebitec.bibigrid.core.model.Configuration;
 
+import static de.unibi.cebitec.bibigrid.core.util.VerboseOutputFilter.V;
+
 /**
- * @author mfriedrichs(at)techfak.uni-bielefeld.de
+ * @author mfriedrichs(at)techfak.uni-bielefeld.de, jkrueger(at)cebitec.uni-bielefeld.de
  */
 @SuppressWarnings({"WeakerAccess", "unused"})
 public class ConfigurationAWS extends Configuration {
@@ -12,7 +14,8 @@ public class ConfigurationAWS extends Configuration {
 
     private double bidPrice;
     private double bidPriceMaster;
-    private boolean publicSlaveIps;
+    private boolean publicSlaveIps = false;
+    private boolean useSpotInstances = false;
 
     public double getBidPrice() {
         return bidPrice;
@@ -36,5 +39,16 @@ public class ConfigurationAWS extends Configuration {
 
     public void setPublicSlaveIps(boolean publicSlaveIps) {
         this.publicSlaveIps = publicSlaveIps;
+    }
+
+    public boolean isUseSpotInstances() {
+        return useSpotInstances;
+    }
+
+    public void setUseSpotInstances(boolean useSpotInstances) {
+        this.useSpotInstances = useSpotInstances;
+        if (useSpotInstances) {
+            LOG.info(V, "Use spot request for all");
+        }
     }
 }
