+++ bibigrid-aws/src/main/java/de/unibi/cebitec/bibigrid/aws/ProviderModuleAWS.java
@@ -1,11 +1,11 @@
 package de.unibi.cebitec.bibigrid.aws;
 
-import de.unibi.cebitec.bibigrid.core.CommandLineValidator;
+import de.unibi.cebitec.bibigrid.core.Validator;
 import de.unibi.cebitec.bibigrid.core.intents.*;
 import de.unibi.cebitec.bibigrid.core.model.*;
 import de.unibi.cebitec.bibigrid.core.model.exceptions.ClientConnectionFailedException;
 import de.unibi.cebitec.bibigrid.core.model.exceptions.ConfigurationException;
-import de.unibi.cebitec.bibigrid.core.util.DefaultPropertiesFile;
+import de.unibi.cebitec.bibigrid.core.util.ConfigurationFile;
 import org.apache.commons.cli.CommandLine;
 
 import java.util.HashMap;
@@ -22,11 +22,11 @@ public class ProviderModuleAWS extends ProviderModule {
     }
 
     @Override
-    public CommandLineValidator getCommandLineValidator(final CommandLine commandLine,
-                                                        final DefaultPropertiesFile defaultPropertiesFile,
-                                                        final IntentMode intentMode)
+    public Validator getCommandLineValidator(final CommandLine commandLine,
+                                             final ConfigurationFile configurationFile,
+                                             final IntentMode intentMode)
             throws ConfigurationException {
-        return new CommandLineValidatorAWS(commandLine, defaultPropertiesFile, intentMode, this);
+        return new ValidatorAWS(commandLine, configurationFile, intentMode, this);
     }
 
     @Override
