+++ bibigrid-core/pom.xml
@@ -15,11 +15,6 @@
     <dependencies>
         <!-- General -->
         <dependency>
-            <groupId>com.github.BiBiServ.bibiserv</groupId>
-            <artifactId>bibiservabstraction</artifactId>
-            <version>2.1.0</version>
-        </dependency>
-        <dependency>
             <groupId>commons-cli</groupId>
             <artifactId>commons-cli</artifactId>
             <version>1.4</version>
