+++ bibigrid-core/src/main/java/de/unibi/cebitec/bibigrid/core/model/ProviderModule.java
@@ -1,11 +1,11 @@
 package de.unibi.cebitec.bibigrid.core.model;
 
-import de.unibi.cebitec.bibigrid.core.CommandLineValidator;
+import de.unibi.cebitec.bibigrid.core.Validator;
 import de.unibi.cebitec.bibigrid.core.intents.*;
 import de.unibi.cebitec.bibigrid.core.model.exceptions.ClientConnectionFailedException;
 import de.unibi.cebitec.bibigrid.core.model.exceptions.ConfigurationException;
 import de.unibi.cebitec.bibigrid.core.model.exceptions.InstanceTypeNotFoundException;
-import de.unibi.cebitec.bibigrid.core.util.DefaultPropertiesFile;
+import de.unibi.cebitec.bibigrid.core.util.ConfigurationFile;
 import org.apache.commons.cli.CommandLine;
 
 import java.util.Collection;
@@ -30,8 +30,8 @@ public abstract class ProviderModule {
      * Get the command line validator implementation for the specified provider, that can handle
      * provider specific parameters.
      */
-    public abstract CommandLineValidator getCommandLineValidator(
-            final CommandLine commandLine, final DefaultPropertiesFile defaultPropertiesFile, final IntentMode intentMode)
+    public abstract Validator getCommandLineValidator(
+            final CommandLine commandLine, final ConfigurationFile configurationFile, final IntentMode intentMode)
             throws ConfigurationException;
 
     public abstract Client getClient(Configuration config) throws ClientConnectionFailedException;
