+++ bibigrid-googlecloud/src/main/java/de/unibi/cebitec/bibigrid/googlecloud/ConfigurationGoogleCloud.java
@@ -3,12 +3,13 @@ package de.unibi.cebitec.bibigrid.googlecloud;
 import de.unibi.cebitec.bibigrid.core.model.Configuration;
 
 /**
- * @author mfriedrichs(at)techfak.uni-bielefeld.de
+ * @author mfriedrichs(at)techfak.uni-bielefeld.de, jkrueger(at)cebitec.uni-bielefeld.de
  */
 @SuppressWarnings({"WeakerAccess", "unused"})
 public class ConfigurationGoogleCloud extends Configuration {
     private String googleProjectId;
     private String googleImageProjectId;
+    private boolean useSpotInstances;
 
     public ConfigurationGoogleCloud() {
     }
@@ -27,5 +28,13 @@ public class ConfigurationGoogleCloud extends Configuration {
 
     public void setGoogleImageProjectId(String googleImageProjectId) {
         this.googleImageProjectId = googleImageProjectId.trim();
+    }
+
+    public boolean isUseSpotInstances() {
+        return useSpotInstances;
+    }
+
+    public void setUseSpotInstances(boolean useSpotInstances) {
+        this.useSpotInstances = useSpotInstances;
     }
 }
