+++ bibigrid-googlecloud/src/main/java/de/unibi/cebitec/bibigrid/googlecloud/ProviderModuleGoogleCloud.java
@@ -3,12 +3,12 @@ package de.unibi.cebitec.bibigrid.googlecloud;
 import com.google.api.services.compute.Compute;
 import com.google.api.services.compute.model.MachineType;
 import com.google.api.services.compute.model.MachineTypesScopedList;
-import de.unibi.cebitec.bibigrid.core.CommandLineValidator;
+import de.unibi.cebitec.bibigrid.core.Validator;
 import de.unibi.cebitec.bibigrid.core.intents.*;
 import de.unibi.cebitec.bibigrid.core.model.*;
 import de.unibi.cebitec.bibigrid.core.model.exceptions.ClientConnectionFailedException;
 import de.unibi.cebitec.bibigrid.core.model.exceptions.ConfigurationException;
-import de.unibi.cebitec.bibigrid.core.util.DefaultPropertiesFile;
+import de.unibi.cebitec.bibigrid.core.util.ConfigurationFile;
 import org.apache.commons.cli.CommandLine;
 
 import java.util.HashMap;
@@ -24,11 +24,11 @@ public class ProviderModuleGoogleCloud extends ProviderModule {
     }
 
     @Override
-    public CommandLineValidator getCommandLineValidator(final CommandLine commandLine,
-                                                        final DefaultPropertiesFile defaultPropertiesFile,
-                                                        final IntentMode intentMode)
+    public Validator getCommandLineValidator(final CommandLine commandLine,
+                                             final ConfigurationFile configurationFile,
+                                             final IntentMode intentMode)
             throws ConfigurationException {
-        return new CommandLineValidatorGoogleCloud(commandLine, defaultPropertiesFile, intentMode, this);
+        return new ValidatorGoogleCloud(commandLine, configurationFile, intentMode, this);
     }
 
     @Override
