+++ bibigrid-main/pom.xml
@@ -89,11 +89,6 @@
     <dependencies>
         <!-- General -->
         <dependency>
-            <groupId>com.github.BiBiServ.bibiserv</groupId>
-            <artifactId>bibiservabstraction</artifactId>
-            <version>2.1.0</version>
-        </dependency>
-        <dependency>
             <groupId>de.unibi.cebitec.bibigrid</groupId>
             <artifactId>bibigrid-core</artifactId>
             <version>2.0.1</version>
